+++ core/GrammarLanguage.ts
@@ -69,6 +69,8 @@ enum GrammarConstants {
   enumFromCellTypes = "enumFromCellTypes", // temporary?
   enum = "enum", // temporary?
   examples = "examples",
+  min = "min",
+  max = "max",
 
   // baseNodeTypes
   baseNodeType = "baseNodeType",
@@ -230,14 +232,14 @@ abstract class GrammarBackedNode extends TreeNode {
     return this.getTopDownArray().filter((node: GrammarBackedNode) => node.getDefinition().getNodeTypeIdFromDefinition() === nodeTypeId)
   }
 
-  getInPlaceCellTypeTree() {
+  toCellTypeTree() {
     return this.getTopDownArray()
       .map(child => child.getIndentation() + child.getLineCellTypes())
       .join("\n")
   }
 
   getParseTable(maxColumnWidth = 40) {
-    const tree = new TreeNode(this.getInPlaceCellTypeTree())
+    const tree = new TreeNode(this.toCellTypeTree())
     return new TreeNode(
       tree.getTopDownArray().map((node, lineNumber) => {
         const sourceNode = this.nodeAtLine(lineNumber)
@@ -346,19 +348,19 @@ abstract class GrammarBackedNode extends TreeNode {
     return usage
   }
 
-  getInPlaceHighlightScopeTree() {
+  toHighlightScopeTree() {
     return this.getTopDownArray<GrammarBackedNode>()
       .map(child => child.getIndentation() + child.getLineHighlightScopes())
       .join("\n")
   }
 
-  getInPlaceCellTypeTreeWithNodeConstructorNames() {
+  toCellTypeTreeWithNodeConstructorNames() {
     return this.getTopDownArray()
       .map(child => child.constructor.name + this.getWordBreakSymbol() + child.getIndentation() + child.getLineCellTypes())
       .join("\n")
   }
 
-  getInPlacePreludeCellTypeTreeWithNodeConstructorNames() {
+  toPreludeCellTypeTreeWithNodeConstructorNames() {
     return this.getTopDownArray<GrammarBackedNode>()
       .map(child => child.constructor.name + this.getWordBreakSymbol() + child.getIndentation() + child.getLineCellPreludeTypes())
       .join("\n")
@@ -384,8 +386,8 @@ abstract class GrammarBackedNode extends TreeNode {
     const treeMTime = this.getLineOrChildrenModifiedTime()
     if (this._cache_programCellTypeStringMTime === treeMTime) return undefined
 
-    this._cache_typeTree = new TreeNode(this.getInPlaceCellTypeTree())
-    this._cache_highlightScopeTree = new TreeNode(this.getInPlaceHighlightScopeTree())
+    this._cache_typeTree = new TreeNode(this.toCellTypeTree())
+    this._cache_highlightScopeTree = new TreeNode(this.toHighlightScopeTree())
     this._cache_programCellTypeStringMTime = treeMTime
   }
 
@@ -599,6 +601,18 @@ abstract class AbstractGrammarBackedCell<T> {
     return this._isCatchAll
   }
 
+  get min() {
+    return this._getCellTypeDefinition().get(GrammarConstants.min) || "0"
+  }
+
+  get max() {
+    return this._getCellTypeDefinition().get(GrammarConstants.max) || "100"
+  }
+
+  get placeholder() {
+    return this._getCellTypeDefinition().get(GrammarConstants.examples) || ""
+  }
+
   abstract getParsed(): T
 
   getHighlightScope(): string | undefined {
@@ -622,14 +636,41 @@ abstract class AbstractGrammarBackedCell<T> {
     })
   }
 
-  synthesizeCell(): string {
+  synthesizeCell(seed = Date.now()): string {
     // todo: cleanup
     const cellDef = this._getCellTypeDefinition()
     const enumOptions = cellDef._getFromExtended(GrammarConstants.enum)
-    return enumOptions ? TreeUtils.getRandomString(1, enumOptions.split(" ")) : this._synthesizeCell()
+    if (enumOptions) return TreeUtils.getRandomString(1, enumOptions.split(" "))
+
+    return this._synthesizeCell(seed)
   }
 
-  abstract _synthesizeCell(): string
+  _getStumpEnumInput(crux: string): string {
+    const cellDef = this._getCellTypeDefinition()
+    const enumOptions = cellDef._getFromExtended(GrammarConstants.enum)
+    if (!enumOptions) return undefined
+    const options = new TreeNode(
+      enumOptions
+        .split(" ")
+        .map(option => `option ${option}`)
+        .join("\n")
+    )
+    return `select
+ name ${crux}
+${options.toString(1)}`
+  }
+
+  _toStumpInput(crux: string): string {
+    // todo: remove
+    const enumInput = this._getStumpEnumInput(crux)
+    if (enumInput) return enumInput
+    // todo: cleanup. We shouldn't have these dual cellType classes.
+    return `input
+ name ${crux}
+ placeholder ${this.placeholder}`
+  }
+
+  abstract _synthesizeCell(seed?: number): string
 
   _getCellTypeDefinition() {
     return this._typeDef
@@ -661,32 +702,6 @@ abstract class AbstractGrammarBackedCell<T> {
   }
 }
 
-class GrammarIntCell extends AbstractGrammarBackedCell<number> {
-  _isValid() {
-    const word = this.getWord()
-    const num = parseInt(word)
-    if (isNaN(num)) return false
-    return num.toString() === word
-  }
-
-  static defaultHighlightScope = "constant.numeric.integer"
-
-  _synthesizeCell() {
-    return TreeUtils.getRandomString(2, "123456789".split(""))
-  }
-
-  getRegexString() {
-    return "\-?[0-9]+"
-  }
-
-  getParsed() {
-    const word = this.getWord()
-    return parseInt(word)
-  }
-
-  static parserFunctionName = "parseInt"
-}
-
 class GrammarBitCell extends AbstractGrammarBackedCell<boolean> {
   _isValid() {
     const word = this.getWord()
@@ -709,7 +724,44 @@ class GrammarBitCell extends AbstractGrammarBackedCell<boolean> {
   }
 }
 
-class GrammarFloatCell extends AbstractGrammarBackedCell<number> {
+abstract class GrammarNumericCell extends AbstractGrammarBackedCell<number> {
+  _toStumpInput(crux: string): string {
+    return `input
+ name ${crux}
+ type number
+ placeholder ${this.placeholder}
+ min ${this.min}
+ max ${this.max}`
+  }
+}
+
+class GrammarIntCell extends GrammarNumericCell {
+  _isValid() {
+    const word = this.getWord()
+    const num = parseInt(word)
+    if (isNaN(num)) return false
+    return num.toString() === word
+  }
+
+  static defaultHighlightScope = "constant.numeric.integer"
+
+  _synthesizeCell(seed) {
+    return TreeUtils.randomUniformInt(parseInt(this.min), parseInt(this.max), seed)
+  }
+
+  getRegexString() {
+    return "\-?[0-9]+"
+  }
+
+  getParsed() {
+    const word = this.getWord()
+    return parseInt(word)
+  }
+
+  static parserFunctionName = "parseInt"
+}
+
+class GrammarFloatCell extends GrammarNumericCell {
   _isValid() {
     const word = this.getWord()
     const num = parseFloat(word)
@@ -718,8 +770,8 @@ class GrammarFloatCell extends AbstractGrammarBackedCell<number> {
 
   static defaultHighlightScope = "constant.numeric.float"
 
-  _synthesizeCell() {
-    return TreeUtils.getRandomString(2, "123456789".split("")) + "." + TreeUtils.getRandomString(2, "0123456789".split(""))
+  _synthesizeCell(seed) {
+    return TreeUtils.randomUniformFloat(parseFloat(this.min), parseFloat(this.max), seed)
   }
 
   getRegexString() {
@@ -1830,6 +1882,14 @@ abstract class AbstractGrammarDefinitionNode extends AbstractExtendibleTreeNode 
     if (enumOptions) return `'^ *(${TreeUtils.escapeRegExp(enumOptions.join("|"))})(?: |$)'`
   }
 
+  toStumpString() {
+    const cellParser = this.getCellParser()
+    const requiredCellTypeIds = cellParser.getRequiredCellTypeIds()
+    const catchAllCellTypeId = cellParser.getCatchAllCellTypeId()
+    return `div
+ label ${this._getCruxIfAny()}`
+  }
+
   // todo: refactor. move some parts to cellParser?
   _toSublimeMatchBlock() {
     const defaultHighlightScope = "source"
@@ -1897,12 +1957,34 @@ ${captures}
     if (ancestorIds.length > 1) return ancestorIds[ancestorIds.length - 2]
   }
 
-  private _generateSimulatedLine(): string {
+  protected _toStumpString() {
+    const crux = this._getCruxIfAny()
+    const cellArray = this.getCellParser()
+      .getCellArray()
+      .filter((item, index) => index) // for now this only works for keyword langs
+    if (!cellArray.length)
+      // todo: remove this! just doing it for now until we refactor getCellArray to handle catchAlls better.
+      return ""
+    const cells = new TreeNode(cellArray.map((cell, index) => cell._toStumpInput(crux)).join("\n"))
+    return `div
+ label ${crux}
+${cells.toString(1)}`
+  }
+
+  toStumpString() {
+    const nodeBreakSymbol = "\n"
+    return this._getInScopeNodeTypeIds()
+      .map(nodeTypeId => this.getNodeTypeDefinitionByNodeTypeId(nodeTypeId)._toStumpString())
+      .filter()
+      .join(nodeBreakSymbol)
+  }
+
+  private _generateSimulatedLine(seed: number): string {
     // todo: generate simulated data from catch all
     const crux = this._getCruxIfAny()
     return this.getCellParser()
       .getCellArray()
-      .map((cell, index) => (!index && crux ? crux : cell.synthesizeCell()))
+      .map((cell, index) => (!index && crux ? crux : cell.synthesizeCell(seed)))
       .join(" ")
   }